+++ typescript/sdk/src/token/nativeTokenMetadata.ts
@@ -1,0 +1,25 @@
+import { ProtocolType } from '@hyperlane-xyz/utils';
+
+import { NativeToken } from '../metadata/chainMetadataTypes.js';
+
+export const PROTOCOL_TO_DEFAULT_NATIVE_TOKEN: Record<
+  ProtocolType,
+  NativeToken
+> = {
+  [ProtocolType.Ethereum]: {
+    decimals: 18,
+    name: 'Ether',
+    symbol: 'ETH',
+  },
+  [ProtocolType.Sealevel]: {
+    decimals: 9,
+    name: 'Solana',
+    symbol: 'SOL',
+  },
+  [ProtocolType.Cosmos]: {
+    decimals: 6,
+    denom: 'uatom',
+    name: 'Atom',
+    symbol: 'ATOM',
+  },
+};